requirements.txt
anthropic
fastapi
uvicorn[standard]
python-dotenv
mcp
